Move sorting classes to `service` package for better organization.
