+++ StellaSort/src/main/java/com/aldebaran/stellasort/service/BubbleSort.java
@@ -1,4 +1,4 @@
-package com.aldebaran.stellasort;
+package com.aldebaran.stellasort.service;
 
 public class BubbleSort {
