+++ StellaSort/src/main/java/com/aldebaran/stellasort/service/QuickSort.java
@@ -1,4 +1,4 @@
-package com.aldebaran.stellasort;
+package com.aldebaran.stellasort.service;
 
 public class QuickSort {
